Closes #1749: Prevent transaction creation on file upload and speed up listing files in zips
